+++ app/src/main/java/com/neocalc/neocalc/history/presentation/HistoryScreen.kt
@@ -4,19 +4,10 @@ import android.content.res.Configuration
 import androidx.compose.foundation.background
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListScope
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
@@ -32,24 +23,18 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.State
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.ehsanmsz.mszprogressindicator.progressindicator.BallClipRotateProgressIndicator
-import com.ehsanmsz.mszprogressindicator.progressindicator.BallPulseProgressIndicator
 import com.ehsanmsz.mszprogressindicator.progressindicator.LineScaleProgressIndicator
 import com.neocalc.neocalc.R
-import com.neocalc.neocalc.history.domain.entities.HistoryItem
+import com.neocalc.neocalc.history.presentation.components.ItemDivider
 import com.neocalc.neocalc.history.domain.entities.HistoryType
 import com.neocalc.neocalc.history.domain.entities.ListState
 import com.neocalc.neocalc.history.presentation.components.InfiniteScrollLazyColumn
@@ -218,40 +203,15 @@ fun HistoryItem(
 
 @Composable
 fun DateItem(date: String, modifier: Modifier = Modifier) {
-	Column(modifier.fillMaxWidth()) {
-
-		// might place this in lazy list
-		Spacer(modifier = Modifier.height(10.dp))
-
-		Text(
-			text = date,
-			modifier = modifier
-				.fillMaxWidth()
-				.padding(horizontal = 16.dp),
-			textAlign = TextAlign.Start,
-			style = MaterialTheme.typography.titleMedium.copy(
-				color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.75f)
-			)
-		)
-	}
-}
-
-@Composable
-fun LazyListScope.ItemDivider(
-	modifier: Modifier = Modifier,
-	dividerColor: Color = Color.DarkGray,
-	dividerHeight: Dp = 2.dp
-) {
-	Box(
+	Text(
+		text = date,
 		modifier = modifier
 			.fillMaxWidth()
-			.height(dividerHeight)
-			.background(dividerColor)
+			.padding(vertical = 10.dp, horizontal = 16.dp),
+		textAlign = TextAlign.Start,
+		style = MaterialTheme.typography.titleMedium.copy(
+			color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.75f)
+		)
 	)
-}
-
-fun LazyListState.hasReachedBottom(buffer: Int = 1): Boolean {
-	val lastVisibleItem = this.layoutInfo.visibleItemsInfo.lastOrNull()
-	return lastVisibleItem?.index != 0 && (lastVisibleItem?.index == this.layoutInfo.totalItemsCount - 3)
 }
